move errors files to folder
