--- controllers/sql-error controller.js
@@ -1,5 +1,0 @@
-exports.sqlErr = (err, req, res, next) => {
-  if (err.code === "22P02") {
-    res.status(400).send({ msg: "invalid input type" });
-  } else next(err);
-};
